+++ deployments/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 import deployments.models as models
+from api.admin_classes import RegionRestrictedAdmin
 
 
 class ERUInline(admin.TabularInline):
@@ -7,13 +8,17 @@ class ERUInline(admin.TabularInline):
     autocomplete_fields = ('deployed_to', 'event',)
 
 
-class ERUOwnerAdmin(admin.ModelAdmin):
+class ERUOwnerAdmin(RegionRestrictedAdmin):
+    country_in = 'national_society_country__in'
+    region_in = 'national_society_country__region__in'
     inlines = [ERUInline]
     autocomplete_fields = ('national_society_country',)
     search_fields = ('national_society_country__name',)
 
 
-class HeopAdmin(admin.ModelAdmin):
+class HeopAdmin(RegionRestrictedAdmin):
+    country_in = 'country__pk__in'
+    region_in = 'region__pk__in'
     search_fields = ('country__name', 'region__name', 'person', 'role',)
 
 
@@ -21,7 +26,9 @@ class FactPersonInline(admin.TabularInline):
     model = models.FactPerson
 
 
-class FactAdmin(admin.ModelAdmin):
+class FactAdmin(RegionRestrictedAdmin):
+    country_in = 'country__pk__in'
+    region_in = 'region__pk__in'
     inlines = [FactPersonInline]
     search_fields = ('country__name', 'region__name',)
 
@@ -30,7 +37,9 @@ class RdrtPersonInline(admin.TabularInline):
     model = models.RdrtPerson
 
 
-class RdrtAdmin(admin.ModelAdmin):
+class RdrtAdmin(RegionRestrictedAdmin):
+    country_in = 'country__pk__in'
+    region_in = 'region__pk__in'
     inlines = [RdrtPersonInline]
     search_fields = ('country__name', 'region__name',)
 
@@ -39,7 +48,9 @@ class PartnerSocietyActivityAdmin(admin.ModelAdmin):
     search_fields = ('activity',)
 
 
-class PartnerSocietyDeploymentAdmin(admin.ModelAdmin):
+class PartnerSocietyDeploymentAdmin(RegionRestrictedAdmin):
+    country_in = 'parent_society__in'
+    region_in = 'parent_society__region__in'
     autocomplete_fields = ('parent_society', 'country_deployed_to', 'district_deployed_to',)
     search_fields = ('activity__activity', 'name', 'role', 'country_deployed_to__name', 'parent_society__name', 'district_deployed_to__name',)