+++ notifications/notification.py
@@ -31,6 +31,7 @@ class SendMail(threading.Thread):
         except SMTPAuthenticationError:
             logger.error('SMTPAuthenticationError')
             logger.error('Cannot send notification')
+            logger.error(str(SMTPAuthenticationError)[:100])
 
 
 def send_notification (subject, recipients, html):